requirements.txt
opencv-python
face-recognition
numpy
Pillow
tk
configparser